[Janos#3] Refactored package structure
