+++ janos-api/src/main/java/org/apache/olingo/odata2/janos/processor/api/data/source/FunctionExecutor.java
@@ -16,7 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  ******************************************************************************/
-package org.apache.olingo.odata2.janos.processor.api.datasource;
+package org.apache.olingo.odata2.janos.processor.api.data.source;
+
+import org.apache.olingo.odata2.janos.processor.api.data.store.DataStoreManager;
 
 /**
  * A FunctionExecutor defines a class which is used to execute a function and
