+++ janos-api/src/main/java/org/apache/olingo/odata2/janos/processor/api/data/store/DataStoreException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.olingo.odata2.janos.processor.api.datasource;
+package org.apache.olingo.odata2.janos.processor.api.data.store;
 
 import org.apache.olingo.odata2.api.exception.ODataApplicationException;
