+++ janos-api/src/main/java/org/apache/olingo/odata2/janos/processor/api/data/store/DataStoreManager.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.apache.olingo.odata2.janos.processor.api.datasource;
+package org.apache.olingo.odata2.janos.processor.api.data.store;
 
 import java.util.Map;
