--- janos-api/src/main/java/org/apache/olingo/odata2/janos/processor/api/datasource/ReadResult.java
@@ -1,16 +1,0 @@
-package org.apache.olingo.odata2.janos.processor.api.datasource;
-
-import java.util.Collection;
-
-/**
- * Created by mibo on 04.09.15.
- */
-public interface ReadResult<T> {
-
-  Collection<T> getResult();
-
-  boolean appliedSkip();
-  boolean appliedTop();
-  boolean appliedOrder();
-  boolean appliedFilter();
-}
