+++ janos-core/src/main/java/org/apache/olingo/odata2/janos/processor/core/data/store/JpaAnnotationDataStore.java
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.olingo.odata2.janos.processor.core.datasource;
+package org.apache.olingo.odata2.janos.processor.core.data.store;
 
 import org.apache.olingo.odata2.api.annotation.edm.EdmKey;
-import org.apache.olingo.odata2.janos.processor.api.datasource.DataStore;
-import org.apache.olingo.odata2.janos.processor.api.datasource.DataStoreException;
-import org.apache.olingo.odata2.janos.processor.api.datasource.ReadOptions;
-import org.apache.olingo.odata2.janos.processor.api.datasource.ReadResult;
+import org.apache.olingo.odata2.janos.processor.api.data.store.DataStore;
+import org.apache.olingo.odata2.janos.processor.api.data.store.DataStoreException;
+import org.apache.olingo.odata2.janos.processor.api.data.ReadOptions;
+import org.apache.olingo.odata2.janos.processor.api.data.ReadResult;
 import org.apache.olingo.odata2.janos.processor.core.util.AnnotationHelper;
 import org.apache.olingo.odata2.janos.processor.core.util.AnnotationRuntimeException;
 
@@ -128,7 +128,7 @@ public class JpaAnnotationDataStore<T> implements DataStore<T> {
   @Override
   public ReadResult<T> read(ReadOptions readOptions) {
     Query query = entityManager.createQuery("SELECT t FROM " + dataTypeClass.getSimpleName() + " t");
-    return GenericReadResult.forResult(query.getResultList()).build();
+    return ReadResult.forResult(query.getResultList()).build();
   }
 
   @Override
