+++ janos-core/src/main/java/org/apache/olingo/odata2/janos/processor/core/rt/JanosServiceBuilderImpl.java
@@ -24,15 +24,15 @@ import org.apache.olingo.odata2.api.exception.ODataException;
 import org.apache.olingo.odata2.api.rt.RuntimeDelegate;
 import org.apache.olingo.odata2.janos.processor.api.JanosService;
 import org.apache.olingo.odata2.janos.processor.api.JanosService.JanosServiceBuilder;
-import org.apache.olingo.odata2.janos.processor.api.datasource.DataSource;
-import org.apache.olingo.odata2.janos.processor.api.datasource.DataStoreManager;
-import org.apache.olingo.odata2.janos.processor.api.datasource.FunctionSource;
-import org.apache.olingo.odata2.janos.processor.api.datasource.ValueAccess;
+import org.apache.olingo.odata2.janos.processor.api.data.source.DataSource;
+import org.apache.olingo.odata2.janos.processor.api.data.store.DataStoreManager;
+import org.apache.olingo.odata2.janos.processor.api.data.source.FunctionSource;
+import org.apache.olingo.odata2.janos.processor.api.data.access.ValueAccess;
 import org.apache.olingo.odata2.janos.processor.core.DataSourceProcessor;
-import org.apache.olingo.odata2.janos.processor.core.datasource.AnnotationDataSource;
-import org.apache.olingo.odata2.janos.processor.core.datasource.AnnotationFunctionSource;
-import org.apache.olingo.odata2.janos.processor.core.datasource.AnnotationValueAccess;
-import org.apache.olingo.odata2.janos.processor.core.datasource.DualDataStoreManager;
+import org.apache.olingo.odata2.janos.processor.core.data.source.AnnotationDataSource;
+import org.apache.olingo.odata2.janos.processor.core.data.source.AnnotationFunctionSource;
+import org.apache.olingo.odata2.janos.processor.core.data.access.AnnotationValueAccess;
+import org.apache.olingo.odata2.janos.processor.core.data.store.DualDataStoreManager;
 import org.apache.olingo.odata2.janos.processor.core.edm.AnnotationEdmProvider;
 
 import java.util.ArrayList;
