+++ janos-core/src/test/java/org/apache/olingo/odata2/janos/processor/core/DataSourceProcessorTest.java
@@ -16,11 +16,11 @@
 package org.apache.olingo.odata2.janos.processor.core;
 
 import org.apache.olingo.odata2.api.exception.ODataException;
-import org.apache.olingo.odata2.janos.processor.api.datasource.DataSource;
-import org.apache.olingo.odata2.janos.processor.api.datasource.DataStoreManager;
-import org.apache.olingo.odata2.janos.processor.api.datasource.ValueAccess;
-import org.apache.olingo.odata2.janos.processor.core.datasource.AnnotationDataSource;
-import org.apache.olingo.odata2.janos.processor.core.datasource.AnnotationValueAccess;
+import org.apache.olingo.odata2.janos.processor.api.data.source.DataSource;
+import org.apache.olingo.odata2.janos.processor.api.data.store.DataStoreManager;
+import org.apache.olingo.odata2.janos.processor.api.data.access.ValueAccess;
+import org.apache.olingo.odata2.janos.processor.core.data.source.AnnotationDataSource;
+import org.apache.olingo.odata2.janos.processor.core.data.access.AnnotationValueAccess;
 import org.apache.olingo.odata2.janos.processor.core.model.Building;
 import org.junit.Assert;
 import org.junit.Test;
