+++ janos-jpa-ref/src/main/java/org/apache/olingo/odata2/janos/processor/ref/jpa/JanosJpaRefServiceFactory.java
@@ -27,9 +27,9 @@ import org.apache.olingo.odata2.api.processor.ODataErrorCallback;
 import org.apache.olingo.odata2.api.processor.ODataErrorContext;
 import org.apache.olingo.odata2.api.processor.ODataResponse;
 import org.apache.olingo.odata2.janos.processor.api.JanosService;
-import org.apache.olingo.odata2.janos.processor.api.datasource.DataStore;
-import org.apache.olingo.odata2.janos.processor.api.datasource.DataStoreException;
-import org.apache.olingo.odata2.janos.processor.core.datasource.JpaAnnotationDataStore;
+import org.apache.olingo.odata2.janos.processor.api.data.store.DataStore;
+import org.apache.olingo.odata2.janos.processor.api.data.store.DataStoreException;
+import org.apache.olingo.odata2.janos.processor.core.data.store.JpaAnnotationDataStore;
 import org.apache.olingo.odata2.janos.processor.ref.jpa.model.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
