+++ janos-ref/src/main/java/org/apache/olingo/odata2/janos/processor/ref/JanosRefServiceFactory.java
@@ -27,8 +27,8 @@ import org.apache.olingo.odata2.api.processor.ODataErrorCallback;
 import org.apache.olingo.odata2.api.processor.ODataErrorContext;
 import org.apache.olingo.odata2.api.processor.ODataResponse;
 import org.apache.olingo.odata2.janos.processor.api.JanosService;
-import org.apache.olingo.odata2.janos.processor.api.datasource.DataStoreException;
-import org.apache.olingo.odata2.janos.processor.core.datasource.InMemoryDataStore;
+import org.apache.olingo.odata2.janos.processor.api.data.store.DataStoreException;
+import org.apache.olingo.odata2.janos.processor.core.data.store.InMemoryDataStore;
 import org.apache.olingo.odata2.janos.processor.ref.model.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
